♻️ Extract BlogArchive component
For re-use on blog category page
