+++ components/BlogArchive.js
@@ -1,0 +1,64 @@
+import { useState, useEffect } from 'react'
+import { fetchData, buildPostsQuery } from 'utils/api'
+import Title from 'components/Title'
+import Wrap from 'components/Wrap'
+import Post from 'components/Post'
+import Link from 'components/Link'
+
+export default function BlogArchive(props) {
+  const [posts, setPosts] = useState(props.posts.edges)
+  const [pageInfo, setPageInfo] = useState(props.posts.pageInfo)
+  const [loading, setLoading] = useState(false)
+  
+  useEffect(() => {
+    if (loading) {
+      const getMore = async () => {
+        const result = await fetchData(`
+          query MorePosts {
+            ${buildPostsQuery(pageInfo.endCursor)}
+          }
+        `).catch(() => setLoading(false))
+
+        if (result) {
+          setLoading(false)
+          setPosts(posts.concat(result.posts.edges))
+          setPageInfo(result.posts.pageInfo)
+        }
+      }
+
+      getMore()
+    }
+  }, [loading])
+
+  return (
+    <Wrap>
+      <div className="lg:flex flex-row-reverse">
+        <div className="flex-1 mb-12 lg:mb-0">
+          <Title>Blog</Title>
+
+          {posts.length ? posts.map(({ node }) => (
+            <Post post={node} key={node.id} />
+          )) : (
+            <div>
+              <p>There are no blog posts yet.</p>
+            </div>
+          )}
+
+          {pageInfo.hasNextPage && (
+            <a className={`${loading ? 'bg-gray-500 text-gray-300' : 'bg-gray-800 hover:bg-gray-700 text-white'} px-3 py-3 font-bold flex justify-center cursor-pointer`} onClick={() => setLoading(true)}>Load more posts</a>
+          )}
+        </div>
+        <div className="lg:w-1/4 lg:mr-16">
+          <h3 className="mb-6">Blog categories</h3>
+          <ul className="list-none border-b">
+            {props.categories.edges.map(({ category }) => (
+              <li key={category.id} className="m-0">
+                <Link href={category.uri} className="flex py-3 px-3 border-t hover:bg-gray-100">{category.name}</Link>
+              </li>
+            ))}
+          </ul>
+        </div>
+      </div>
+    </Wrap>
+  )
+}
